+++ src/main/java/com/chrisenoch/col/CostOfLiving/service/CostOfLivingService.java
@@ -1,6 +1,6 @@
 package com.chrisenoch.col.CostOfLiving.service;
 
-import java.time.OffsetDateTime;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,11 +13,11 @@ public interface CostOfLivingService {
 	
 	public List<COLIndex> findColIndexes(String theDate);
 	
-	public COLResults calculateEquivalentSalary(float amount, String base, String code);
+	public COLResults calculateEquivalentSalary(BigDecimal amount, String base, String code);
 	
 	public Optional<List<COLIndex>> findColIndexesByCountry(String country);
 	
-	public List<COLResults> calculateEquivalentSalaryByCountry(float amount, COLIndex colIndex,  String country);
+	public List<COLResults> calculateEquivalentSalaryByCountry(BigDecimal amount, COLIndex colIndex,  String country);
 	
 	public Optional<COLIndex> findByCity(String city);
